+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 19.1.6
 - you can now **add new bookmarks** with a button in the **bookmark popup**
     - you need QOwnNotes 19.1.6 or higher for this feature
+- you can now **open all visible bookmarks** in new tabs with a button in the **bookmark popup**
 
 ## 19.1.5
 - the previously entered **search text** in the **bookmark popup** is now **automatically selected and focused**,